# swagger.yaml
openapi: "3.0.3"
info:
  title: AgriWizard — Smart Greenhouse Management API
  description: |
    Complete REST API for the AgriWizard microservice ecosystem.
    All protected endpoints require a Bearer JWT token obtained from the IAM service `/login` endpoint.

    ## Services
    | Service | Port | Description |
    |---------|------|-------------|
    | IAM Service | 8081 | Authentication & RBAC |
    | Hardware Service | 8082 | IoT device management & MQTT control |
    | Analytics Service | 8083 | Threshold rules & decision logic |
    | Weather Service | 8084 | External weather intelligence |
    | Traefik Gateway | 8080 | Unified entry point (all routes via `/api/v1/...`) |

    ## Authentication Flow
    1. `POST /api/v1/iam/register` — Create an account
    2. `POST /api/v1/iam/login` — Obtain JWT token
    3. Include `Authorization: Bearer <token>` on all subsequent requests
  version: "1.0.0"
  contact:
    name: AgriWizard Team
    email: support@agriwizard.io
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development (via Traefik gateway)
  - url: http://localhost:8081
    description: IAM Service (direct)
  - url: http://localhost:8082
    description: Hardware Service (direct)
  - url: http://localhost:8083
    description: Analytics Service (direct)
  - url: http://localhost:8084
    description: Weather Service (direct)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from `/api/v1/iam/login`
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: invalid_request
        message:
          type: string
          example: field 'email' is required
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: object

    # ── IAM ─────────────────────────────────────────────────────────
...
    CreateEquipmentRequest:
      type: object
      required: [serial, name, supported_operations]
      properties:
        serial:
          type: string
          example: pump_main_01
        name:
          type: string
          example: Main Water Pump
        supported_operations:
          type: array
          items:
            type: string
          example: [ON, OFF, REVERSE]
        api_url:
          type: string
          example: http://192.168.1.10/control
    Equipment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        serial:
          type: string
          example: pump_main_01
        name:
          type: string
        supported_operations:
          type: array
          items:
            type: string
        mqtt_topic:
          type: string
          example: agriwizard/equipment/pump_main_01/command
        api_url:
          type: string
        current_status:
          type: string
          enum: [ON, OFF, LOCKED, DISABLED]
        created_at:
          type: string
          format: date-time
    ControlCommand:
      type: object
      required: [operation]
      properties:
        operation:
          type: string
          example: ON
        payload:
          type: object
          additionalProperties: true
    CreateSensorRequest:
      type: object
      required: [serial, name, parameter_ids]
      properties:
        serial:
          type: string
          example: soil_probe_zone_a
        name:
          type: string
          example: Zone A Soil Probe
        parameter_ids:
          type: array
          items:
            type: string
          example: [soil_moisture_pct, soil_temp_c]
        api_url:
          type: string
        update_frequency_seconds:
          type: integer
          default: 60
...
    CreateParameterRequest:
      type: object
      required: [id, unit]
      properties:
        id:
          type: string
          example: soil_moisture_pct
        unit:
          type: string
          example: "%"
        description:
          type: string
          example: Volumetric soil moisture percentage
    TelemetryPayload:
      type: object
      required: [sensor_id, readings]
      properties:
        sensor_id:
          type: string
          example: soil_probe_zone_a
        readings:
          type: array
          items:
            type: object
            properties:
              parameter_id:
                type: string
                example: soil_moisture_pct
              value:
                type: number
                format: double
                example: 42.5
        timestamp:
          type: string
          format: date-time

    # ── Analytics ─────────────────────────────────────────────────────
...
paths:
...
  /api/v1/hardware/equipments:
    post:
      summary: Register a new equipment controller
      tags: [Hardware — Equipment]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEquipmentRequest"
      responses:
        "201":
          description: Equipment registered, MQTT topic assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      summary: List all registered equipment
      tags: [Hardware — Equipment]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Equipment list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Equipment"
  /api/v1/hardware/equipments/{id}:
    put:
      summary: Update an existing equipment controller
      tags: [Hardware — Equipment]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEquipmentRequest"
      responses:
        "200":
          description: Equipment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete an equipment controller
      tags: [Hardware — Equipment]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Equipment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/hardware/control/{id}:
    post:
      summary: Dispatch a command to equipment via MQTT
      tags: [Hardware — Equipment]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ControlCommand"
      responses:
        "200":
          description: Command dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Equipment is LOCKED or DISABLED
  /api/v1/hardware/sensors:
    post:
      summary: Provision a new sensor device
      tags: [Hardware — Sensors]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSensorRequest"
      responses:
        "201":
          description: Sensor provisioned, MQTT topic assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    get:
      summary: Get all sensors with their parameter definitions
      tags: [Hardware — Sensors]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Sensor list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /api/v1/hardware/sensors/{id}:
    put:
      summary: Update an existing sensor device
      tags: [Hardware — Sensors]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSensorRequest"
      responses:
        "200":
          description: Sensor updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a sensor device
      tags: [Hardware — Sensors]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Sensor deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/hardware/parameters:
    post:
      summary: Define a new measurable parameter type
      tags: [Hardware — Parameters]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateParameterRequest"
      responses:
        "201":
          description: Parameter defined
    get:
      summary: List all parameter definitions
      tags: [Hardware — Parameters]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Parameter list
  /api/v1/hardware/telemetry:
    post:
      summary: Ingest telemetry via REST (alternative to MQTT)
      tags: [Hardware — Telemetry]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelemetryPayload"
      responses:
        "201":
          description: Telemetry stored and forwarded to analytics

  # ══════════════════════════════════════════════════════════════════
  #  Analytics Service
  # ══════════════════════════════════════════════════════════════════
...